feat: 게시판 형태 공용 레이아웃
