+++ src/App.js
@@ -7,7 +7,6 @@ import Login from './pages/login/Login';
 import Register from './pages/login/Register';
 import NotebookList from './pages/notebook/NotebookList';
 import QuestionList from './pages/question/QuestionList';
-import QuestionListDetail from './pages/question/QuestionListDetail'; 
 
 function App() {
   return (
@@ -24,7 +23,6 @@ function App() {
             </Route>
             <Route path = "question">
               <Route path='list' element={<QuestionList />} />
-              <Route path='detail' element={<QuestionListDetail />} />
             </Route>
           </Route>
         </Routes>
