--- src/components/notebook/Banner.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import { NotebookBanner } from '../../styles/notebook/NotebookList-styled';
-
-const Banner = () => {
-  return <NotebookBanner>노트북 대여</NotebookBanner>;
-};
-
-export default Banner;
